add nacos annotation demo
