+++ README.md
@@ -121,7 +121,7 @@ public class Application {
 
 ```bash
 # mvn clean package
-# mvn -Djava.net.preferIPv4Stack=true -Dexec.mainClass=org.apache.dubbo.demo.provider.Application exec:java
+# mvn -Djava.net.preferIPv4Stack=true -Dexec.mainClass=org.apache.dubbo.demo.provider.NacosProviderApplication exec:java
 ```
 
 ### Call remote service in consumer
@@ -150,7 +150,7 @@ public class Application {
 
 ```bash
 # mvn clean package
-# mvn -Djava.net.preferIPv4Stack=true -Dexec.mainClass=org.apache.dubbo.demo.consumer.Application exec:java
+# mvn -Djava.net.preferIPv4Stack=true -Dexec.mainClass=org.apache.dubbo.demo.consumer.NacosConsumerApplication exec:java
 ```
 
 The consumer will print out `Hello world` on the screen.
