+++ dubbo-demo/dubbo-demo-annotation-nacos/dubbo-demo-annotation-nacos-consumer/src/main/java/org/apache/dubbo/demo/consumer/NacosConsumerApplication.java
@@ -1,0 +1,42 @@
+/*
+ * Licensed to the Apache Software Foundation (ASF) under one
+ * or more contributor license agreements.  See the NOTICE file
+ * distributed with this work for additional information
+ * regarding copyright ownership.  The ASF licenses this file
+ * to you under the Apache License, Version 2.0 (the
+ * "License"); you may not use this file except in compliance
+ * with the License.  You may obtain a copy of the License at
+ *
+ *   http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing,
+ * software distributed under the License is distributed on an
+ * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+ * KIND, either express or implied.  See the License for the
+ * specific language governing permissions and limitations
+ * under the License.
+ */
+package org.apache.dubbo.demo.consumer;
+
+import org.apache.dubbo.config.spring.context.annotation.EnableDubbo;
+import org.apache.dubbo.demo.DemoService;
+import org.apache.dubbo.demo.consumer.comp.DemoServiceComponent;
+
+import org.springframework.context.annotation.AnnotationConfigApplicationContext;
+import org.springframework.context.annotation.ComponentScan;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.context.annotation.PropertySource;
+
+@Configuration
+@EnableDubbo(scanBasePackages = "org.apache.dubbo.demo.consumer.comp")
+@PropertySource("classpath:/spring/dubbo-consumer.properties")
+@ComponentScan(value = {"org.apache.dubbo.demo.consumer.comp"})
+public class NacosConsumerApplication {
+    public static void main(String[] args) {
+        AnnotationConfigApplicationContext context = new AnnotationConfigApplicationContext(NacosConsumerApplication.class);
+        context.start();
+        DemoService service = context.getBean("demoServiceComponent", DemoServiceComponent.class);
+        String hello = service.sayHello("world");
+        System.out.println("result :" + hello);
+    }
+}
